wiegand: Add Wiegand GPIO support
This class adds support for Wiegand GPIO readers. Thusfar it has been
tested with multiple EM41XX based readers.  The next steps will be to
validate it with other units (e.g. HID Proxkey based readers, etc).

To use this define a badge reader as follows in your INI file where the
second argument is the GPIO pin which the Wiegand D0 pin is connected to
and the third argument is the GPIO pin which the Wiegand D1 pin is
connected to:

```
badge_reader = WiegandGPIOReader:3:5
```

The callback for this device returns a 4 byte (8 character), zero padded
hexidecimal string capable of encoding up to 32 bits.  Most commodity
EM41XX readers can present either 26 or 34 bits of data _with the first
and last bit being used for parity_, resulting in 24/32 usable bits of
data respectively.  This may be increased in the future to support 40
bit readers, etc.