WIP: Door authorization example
This is a basic example of using the WiegandGPIOReader class to provide
access control to a single door.  This allows for users to populate a
whitelist and then query the whitelist in a mechanism similar to the
two_button.py example.

This example still needs tests, etc before it should be pushed upstream.