Fix pin numbers used in tests
The tests on the project currently fail on the Raspberry Pi due to a
combination of using GPIO.setmode(BOARD), pin numbers which do not
correspond to GPIO pins in testing, and the number of fake pins
instantiated during setup.

The previous pins used were:
  - 1  (3.3V Power)
  - 2  (5V Power)
  - 3  (SDA)
  - 4  (5V Power)
  - 5  (SCL)
  - 9  (Ground)

This commit moves the pins for testing to the following:
  - 7
  - 11
  - 12
  - 13
  - 15
  - 16

Wherever possible the following mapping was used:
  - 1 -> 11
  - 2 -> 12
  - 3 -> 15
  - 4 -> 16
  - 5 -> 13
  - 9 -> 7

With these changes all tests pass on a Raspberry Pi 3 Model B (a32082)
in both Python version 2.7.16 and Python version 3.7.3

This commit is based off of the modified tests in #30 and requires that
PR to be merged first.